+++ application/ui/src/features/dataset/import-export/import-export.component.tsx
@@ -4,7 +4,8 @@
 import { Button, Item, Key, Menu, MenuTrigger } from '@geti/ui';
 import { useOverlayTriggerState } from '@react-stately/overlays';
 
-import { ExportDataset } from './export-dataset/export-dataset.component';
+import { ExportDatasetConfig } from '../../../components/export-dataset-config-dialog/export-dataset-config.component';
+import { DatasetStatistics } from './export-dataset/dataset-statistics/dataset-statistics.component';
 import { ImportDataset } from './import-dataset/import-dataset.component';
 
 export const ImportExport = () => {
@@ -33,8 +34,8 @@ export const ImportExport = () => {
                 </Menu>
             </MenuTrigger>
 
-            <ExportDataset dialogState={exportDialogState} />
             <ImportDataset dialogState={importDialogState} />
+            <ExportDatasetConfig datasetId={null} dialogState={exportDialogState} statistics={<DatasetStatistics />} />
         </>
     );
 };
