+++ application/ui/src/hooks/use-export-dataset-job-action.hook.tsx
@@ -5,18 +5,20 @@ import { useActionState } from 'react';
 
 import { isString } from 'lodash-es';
 
-import { $api } from '../../../../../api/client';
-import { useExportDataset } from '../../../../../hooks/localStorage/use-export-dataset.hook';
-import { useProjectIdentifier } from '../../../../../hooks/use-project-identifier.hook';
+import { $api } from '../api/client';
+import { useExportDataset } from './localStorage/use-export-dataset.hook';
+import { useProjectIdentifier } from './use-project-identifier.hook';
 
 type FormValues = {
     labels: string[];
     export_format: string;
+    dataset_id: string | null;
     include_unannotated: boolean;
 };
 
 const initialState: FormValues = {
     labels: [],
+    dataset_id: null,
     export_format: 'geti',
     include_unannotated: false,
 };
@@ -31,16 +33,19 @@ export const useExportDatasetJobAction = ({ onSuccess }: useExportDatasetJobActi
     const exportJobMutation = $api.useMutation('post', '/api/jobs');
 
     return useActionState<FormValues, FormData>(async (_prevState, formData) => {
+        const datasetId = String(formData.get('dataset_id'));
+
         const options: FormValues = {
-            export_format: String(formData.get('export_format')),
             labels: formData.getAll('labels').filter(isString),
+            export_format: String(formData.get('export_format')),
             include_unannotated: formData.get('include_unannotated') === 'on',
+            dataset_id: datasetId !== 'null' ? datasetId : null,
         };
 
         const { job_id } = await exportJobMutation.mutateAsync({
             body: {
                 project_id: projectId,
-                dataset_id: null,
+                dataset_id: options.dataset_id,
                 job_type: 'export_dataset',
                 parameters: {
                     export_format: options.export_format,
@@ -54,6 +59,7 @@ export const useExportDatasetJobAction = ({ onSuccess }: useExportDatasetJobActi
 
         addLsExportId(job_id);
         onSuccess();
+        console.log('Export dataset job options:', options);
 
         return options;
     }, initialState);
