+++ application/ui/src/hooks/use-get-dataset-revision-items.hook.ts
@@ -25,7 +25,7 @@ export const useGetDatasetRevisionItems = ({ datasetRevisionId, subset }: UseGet
         '/api/projects/{project_id}/dataset_revisions/{dataset_revision_id}/items',
         {
             params: {
-                query,
+                query: { limit: 1, annotation_status: 'reviewed' },
                 path: { project_id, dataset_revision_id: datasetRevisionId },
             },
         },
